3. Membuat database, Menambahkan data menggunakan seeder, Menampilkan data

+++ database/migrations/2022_06_30_005445_create_employees_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             // AI (auto increment)
             $table->string('nama');
             $table->enum('jeniskelamin',['cowo','cewe']);
-            $table->integer('notelpon');
+            $table->bigInteger('notelpon');
             $table->timestamps();
         });
     }
